+++ packages/forms-web-app/src/dynamic-forms/dynamic-components/multi-file-upload/question.js
@@ -1,29 +1,10 @@
-const {
-	isObjectWithUploadedFiles,
-	getValidFiles,
-	removeFilesV2
-} = require('../../../lib/multi-file-upload-helpers');
-const { createDocument } = require('../../../lib/documents-api-wrapper');
-const { mapMultiFileDocumentToSavedDocument } = require('../../../mappers/document-mapper');
-const {
-	utils: { conjoinedPromises }
-} = require('@pins/common');
 const Question = require('../../question');
-const { JOURNEY_TYPES } = require('@pins/common/src/dynamic-forms/journey-types');
 
 /**
  * @typedef {import('../../journey').Journey} Journey
  * @typedef {import('../../journey-response').JourneyResponse} JourneyResponse
  * @typedef {import('../../section').Section} Section
  * @typedef {import('../../question').QuestionViewModel} QuestionViewModel
- */
-
-/**
- * @typedef {Array<{ id: string; type: string; originalFileName: string; storageId: string }>} UploadedFiles
- */
-
-/**
- * @typedef {QuestionViewModel & UploadedFiles} UploadQuestionViewModel
  */
 
 /**
@@ -48,7 +29,7 @@ class MultiFileUploadQuestion extends Question {
 	 * @param {string} [params.html]
 	 * @param {Array<import('../../question').BaseValidator>} [params.validators]
 	 *
-	 * @param {Array<Function>} [methodOverrides]
+	 * @param {Record<string, Function>} [methodOverrides]
 	 */
 	constructor(
 		{ title, question, fieldName, url, pageTitle, description, documentType, validators, html },
@@ -94,84 +75,6 @@ class MultiFileUploadQuestion extends Question {
 		}
 
 		return viewModel;
-	}
-
-	/**
-	 * Save the answer to the question
-	 * @param {import('express').Request} req
-	 * @param {import('express').Response} res
-	 * @param {Journey} journey
-	 * @param {Section} section
-	 * @param {JourneyResponse} journeyResponse
-	 * @returns {Promise<void>}
-	 */
-	async saveAction(req, res, journey, section, journeyResponse) {
-		// save
-		const { uploadedFiles } = await this.getDataToSave(req, journeyResponse);
-		await Promise.all(
-			uploadedFiles.map((file) => {
-				if (!file) return;
-
-				const data = {
-					...file,
-					type: this.documentType.name
-				};
-
-				return this.uploadDocuments(
-					req.appealsApiClient,
-					journeyResponse.referenceId,
-					journeyResponse.journeyId,
-					data
-				);
-			})
-		);
-		const responseToSave = {
-			answers: {
-				[this.fieldName]: true
-			}
-		};
-		await this.saveResponseToDB(req.appealsApiClient, journey.response, responseToSave);
-
-		// check for saving errors
-		const saveViewModel = this.checkForSavingErrors(req, section, journey);
-		if (saveViewModel) {
-			return this.renderAction(res, saveViewModel);
-		}
-
-		// move to the next question
-		return this.handleNextQuestion(res, journey, section.segment, this.fieldName);
-	}
-
-	// Belongs to save
-	// @ts-ignore
-	uploadDocuments(apiClient, referenceId, journeyId, data) {
-		if ([JOURNEY_TYPES.HAS_QUESTIONNAIRE, JOURNEY_TYPES.S78_QUESTIONNAIRE].includes(journeyId)) {
-			return apiClient.postLPASubmissionDocumentUpload(referenceId, data);
-		} else if ([JOURNEY_TYPES.HAS_APPEAL_FORM, JOURNEY_TYPES.S78_APPEAL_FORM].includes(journeyId)) {
-			return apiClient.postAppellantSubmissionDocumentUpload(referenceId, data);
-		} else if ([JOURNEY_TYPES.S78_LPA_STATEMENT].includes(journeyId)) {
-			return apiClient.postLPAStatementDocumentUpload(referenceId, data);
-		} else if ([JOURNEY_TYPES.S78_APPELLANT_FINAL_COMMENTS].includes(journeyId)) {
-			return apiClient.postAppellantFinalCommentDocumentUpload(referenceId, data);
-		}
-	}
-	// Belongs to save
-	// @ts-ignore
-	removeDocuments(apiClient, journeyId) {
-		// @ts-ignore
-		return (submissionId, documentId) => {
-			if ([JOURNEY_TYPES.HAS_QUESTIONNAIRE, JOURNEY_TYPES.S78_QUESTIONNAIRE].includes(journeyId)) {
-				return apiClient.deleteLPASubmissionDocumentUpload(submissionId, documentId);
-			} else if (
-				[JOURNEY_TYPES.HAS_APPEAL_FORM, JOURNEY_TYPES.S78_APPEAL_FORM].includes(journeyId)
-			) {
-				return apiClient.deleteAppellantSubmissionDocumentUpload(submissionId, documentId);
-			} else if ([JOURNEY_TYPES.S78_LPA_STATEMENT].includes(journeyId)) {
-				return apiClient.deleteLPAStatementDocumentUpload(submissionId, documentId);
-			} else if ([JOURNEY_TYPES.S78_APPELLANT_FINAL_COMMENTS].includes(journeyId)) {
-				return apiClient.deleteAppellantFinalCommentDocumentUpload(submissionId, documentId);
-			}
-		};
 	}
 
 	/**
@@ -224,46 +127,6 @@ class MultiFileUploadQuestion extends Question {
 			visuallyHiddenText: this.question
 		};
 		return action;
-	}
-
-	// probably belongs to save, definitely shouldn't live here
-	// @ts-ignore
-	async saveFilesToBlobStorage(files, journeyResponse) {
-		const resolutions = await conjoinedPromises(files, (file) =>
-			createDocument(
-				{
-					id: this.#generateDocumentSubmissionId(journeyResponse),
-					referenceNumber: journeyResponse.referenceId
-				},
-				file,
-				file.name,
-				this.documentType.name
-			)
-		);
-
-		const result = Array.from(resolutions).map(([file, document]) =>
-			mapMultiFileDocumentToSavedDocument(document, document?.name, file.name)
-		);
-
-		return result;
-	}
-
-	/**
-	 * @param {JourneyResponse} journeyResponse
-	 * @returns {string}
-	 */
-	#generateDocumentSubmissionId(journeyResponse) {
-		return `${journeyResponse.journeyId}:${encodeURIComponent(
-			this.#sanitiseReferenceId(journeyResponse.referenceId)
-		)}`;
-	}
-
-	/**
-	 * @param {string} referenceId
-	 * @returns {string}
-	 */
-	#sanitiseReferenceId(referenceId) {
-		return referenceId.replaceAll('/', '_');
 	}
 
 	/**
