+++ packages/forms-web-app/src/journeys/question-overrides/multi-file-upload.js
@@ -1,13 +1,19 @@
-/**
- * @typedef {import('../../dynamic-forms/journey-response').JourneyResponse} JourneyResponse
- * @typedef {import('../../dynamic-forms/dynamic-components/multi-file-upload/question')} MultiFileUploadQuestion
- */
-
+const { createDocument } = require('#lib/documents-api-wrapper');
 const {
 	isObjectWithUploadedFiles,
 	removeFilesV2,
-	getValidFiles
+	getValidFiles,
+	generateDocumentSubmissionId
 } = require('#lib/multi-file-upload-helpers');
+const { conjoinedPromises } = require('@pins/common/src/utils');
+const { mapMultiFileDocumentToSavedDocument } = require('../../mappers/document-mapper');
+const { JOURNEY_TYPES } = require('@pins/common/src/dynamic-forms/journey-types');
+
+/**
+ * @typedef {import('../../dynamic-forms/journey').Journey} Journey
+ * @typedef {import('../../dynamic-forms/journey-response').JourneyResponse} JourneyResponse
+ * @typedef {import('../../dynamic-forms/section').Section} Section * @typedef {import('../../dynamic-forms/dynamic-components/multi-file-upload/question')} MultiFileUploadQuestion
+ */
 
 /**
  * returns the data to send to the DB
@@ -33,7 +39,7 @@ async function getDataToSave(req, journeyResponse) {
 			removedFiles,
 			journeyResponse.referenceId,
 			`${journeyResponse.journeyId}:${encodedReferenceId}`,
-			this.removeDocuments(req.appealsApiClient, journeyResponse.journeyId)
+			removeDocuments(req.appealsApiClient, journeyResponse.journeyId)
 		);
 
 		// adds validation errors to be checked after
@@ -57,7 +63,7 @@ async function getDataToSave(req, journeyResponse) {
 		return [relevantRequestFiles];
 	})();
 	const validFiles = getValidFiles(errors, filesToUpload);
-	const uploadedFiles = await this.saveFilesToBlobStorage(validFiles, journeyResponse);
+	const uploadedFiles = await saveFilesToBlobStorage.call(this, validFiles, journeyResponse);
 
 	const journeyFiles = {
 		answers: {
@@ -94,6 +100,156 @@ async function getDataToSave(req, journeyResponse) {
 	};
 }
 
+/**
+ * @this {MultiFileUploadQuestion}
+ * @param {import('#lib/multi-file-upload-helpers').File[]} files
+ * @param {JourneyResponse} journeyResponse - current journey response, modified with the new answers
+ * @returns {Promise<{
+ *   message: {
+ *     text: string;
+ *   };
+ *   id: string;
+ *   name: string;
+ *   fileName: string;
+ *   originalFileName: string;
+ *   location: string;
+ *   size: string;
+ * }[]>}
+ */
+async function saveFilesToBlobStorage(files, journeyResponse) {
+	const resolutions = await conjoinedPromises(files, (file) =>
+		createDocument(
+			{
+				id: generateDocumentSubmissionId(journeyResponse),
+				referenceNumber: journeyResponse.referenceId
+			},
+			file,
+			file.name,
+			this.documentType.name
+		)
+	);
+
+	const result = Array.from(resolutions).map(([file, document]) =>
+		mapMultiFileDocumentToSavedDocument(document, document?.name, file.name)
+	);
+
+	return result;
+}
+
+/**
+ * @param {import('@pins/common/src/client/appeals-api-client').AppealsApiClient} apiClient
+ * @param {string} referenceId
+ * @param {string} journeyId
+ * @param {any} data
+ * @returns {Promise<any> | void}
+ */
+function uploadDocuments(apiClient, referenceId, journeyId, data) {
+	switch (journeyId) {
+		case JOURNEY_TYPES.HAS_QUESTIONNAIRE: {
+			return apiClient.postLPASubmissionDocumentUpload(referenceId, data);
+		}
+		case JOURNEY_TYPES.HAS_APPEAL_FORM: {
+			return apiClient.postAppellantSubmissionDocumentUpload(referenceId, data);
+		}
+		case JOURNEY_TYPES.S78_QUESTIONNAIRE: {
+			return apiClient.postLPASubmissionDocumentUpload(referenceId, data);
+		}
+		case JOURNEY_TYPES.S78_APPEAL_FORM: {
+			return apiClient.postAppellantSubmissionDocumentUpload(referenceId, data);
+		}
+		case JOURNEY_TYPES.S78_LPA_STATEMENT: {
+			return apiClient.postLPAStatementDocumentUpload(referenceId, data);
+		}
+		case JOURNEY_TYPES.S78_APPELLANT_FINAL_COMMENTS: {
+			return apiClient.postAppellantFinalCommentDocumentUpload(referenceId, data);
+		}
+		default:
+			throw new Error('Unrecognised journey type');
+	}
+}
+
+/**
+ * @param {import('@pins/common/src/client/appeals-api-client').AppealsApiClient} apiClient
+ * @param {string} journeyId
+ * @returns {(submissionId: string, documentId: string) => Promise<any>}
+ */
+function removeDocuments(apiClient, journeyId) {
+	return (submissionId, documentId) => {
+		switch (journeyId) {
+			case JOURNEY_TYPES.HAS_QUESTIONNAIRE: {
+				return apiClient.deleteLPASubmissionDocumentUpload(submissionId, documentId);
+			}
+			case JOURNEY_TYPES.HAS_APPEAL_FORM: {
+				return apiClient.deleteAppellantSubmissionDocumentUpload(submissionId, documentId);
+			}
+			case JOURNEY_TYPES.S78_QUESTIONNAIRE: {
+				return apiClient.deleteLPASubmissionDocumentUpload(submissionId, documentId);
+			}
+			case JOURNEY_TYPES.S78_APPEAL_FORM: {
+				return apiClient.deleteAppellantSubmissionDocumentUpload(submissionId, documentId);
+			}
+			case JOURNEY_TYPES.S78_LPA_STATEMENT: {
+				return apiClient.deleteLPAStatementDocumentUpload(submissionId, documentId);
+			}
+			case JOURNEY_TYPES.S78_APPELLANT_FINAL_COMMENTS: {
+				return apiClient.deleteAppellantFinalCommentDocumentUpload(submissionId, documentId);
+			}
+			default:
+				throw new Error('Unrecognised journey type');
+		}
+	};
+}
+
+/**
+ * Save the answer to the question
+ * @this {MultiFileUploadQuestion}
+ * @param {import('express').Request} req
+ * @param {import('express').Response} res
+ * @param {Journey} journey
+ * @param {Section} section
+ * @param {JourneyResponse} journeyResponse
+ * @returns {Promise<void>}
+ */
+async function saveAction(req, res, journey, section, journeyResponse) {
+	// save
+	// it would be nice to make this conform to the correct type from the Question method
+	const { uploadedFiles } = await getDataToSave.call(this, req, journeyResponse);
+	await Promise.all(
+		uploadedFiles.map((file) => {
+			if (!file) return;
+
+			const data = {
+				...file,
+				type: this.documentType.name
+			};
+
+			return uploadDocuments(
+				req.appealsApiClient,
+				journeyResponse.referenceId,
+				journeyResponse.journeyId,
+				data
+			);
+		})
+	);
+	const responseToSave = {
+		answers: {
+			[this.fieldName]: true
+		}
+	};
+	await this.saveResponseToDB(req.appealsApiClient, journey.response, responseToSave);
+
+	// check for saving errors
+	const saveViewModel = this.checkForSavingErrors(req, section, journey);
+	if (saveViewModel) {
+		return this.renderAction(res, saveViewModel);
+	}
+
+	// move to the next question
+	return this.handleNextQuestion(res, journey, section.segment, this.fieldName);
+}
+
 module.exports = {
-	getDataToSave
+	getDataToSave,
+	saveFilesToBlobStorage,
+	saveAction
 };
